% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maize.kaeppler.expression}
\alias{maize.kaeppler.expression}
\title{Expression data describing in the paper published by Kaeppler 2015 PMID: 27898762}
\format{A data frame with 32193 rows and 80 variables.}
\usage{
data(maize.kaeppler.expression)
}
\description{
This expression data is mapped to the B73 v2 FGS.  It is available as FPKM values and includes left/right gene positions.
Replicates have been merged by taking the average of the 3 (except for one sample with only 2 replicates).
Tracking IDs have been replaced with the tissue sample used and are under the column "Sample".
}
\keyword{datasets}
