% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maize.kaeppler.expression.replicate}
\alias{maize.kaeppler.expression.replicate}
\title{Expression data describing in the paper published by Kaeppler 2015 PMID: 27898762}
\format{A data frame with 32193 rows and 80 variables.}
\usage{
data(maize.kaeppler.expression.replicate)
}
\description{
This expression data is mapped to the B73 v2 FGS.  It is available as FPKM values and includes left/right gene positions.
}
\keyword{datasets}
