% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maize.walley.abundance}
\alias{maize.walley.abundance}
\title{Protein abundance data describing in the paper published by Walley 2016 PMID: 27540173}
\format{A data frame in long format.}
\usage{
data(maize.walley.abundance)
}
\description{
This expression data is mapped to the B73 v2 FGS.  It is available as dNSAF values.
Replicates have been merged by taking the average of the 3 (except for one sample with only 2 replicates).
Tracking IDs have been replaced with the tissue sample used and are under the column "Sample".
}
\keyword{datasets}
