% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maize.walley.expression}
\alias{maize.walley.expression}
\title{Expression data describing in the paper published by Walley 2016 PMID: 27540173}
\format{A data frame in long format.}
\usage{
data(maize.walley.expression)
}
\description{
This expression data is mapped to the B73 v2 WGS.  It is available as FPKM values.
Replicates have been merged by taking the average of the 3 (except for one sample with only 2 replicates).
Tracking ID's have been replaced with the tissue sample used and are under the column "Sample".
}
\keyword{datasets}
