% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maize.walley.expression.replicate}
\alias{maize.walley.expression.replicate}
\title{Expression data describing in the paper published by Walley 2016 PMID: 27540173}
\format{A data frame with 32175 rows and 69 variables.}
\usage{
data(maize.walley.expression.replicate)
}
\description{
This expression data is mapped to the B73 v2 WGS.  It is available as FPKM values.
}
\keyword{datasets}
