% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maize.walley.v4mapped.expression.replicate}
\alias{maize.walley.v4mapped.expression.replicate}
\title{Expression data describing in the paper published by Walley 2016 PMID: 27540173}
\format{A data frame with 32175 rows and 69 variables.}
\usage{
data(maize.walley.v4mapped.expression.replicate)
}
\description{
This expression data was remapped to the B73 v4.37 assembly using STAR.  It is available as FPKM values.
}
\keyword{datasets}
